authentication bug fixed , and user function working

+++ src/dbHelpers.ts
@@ -3,10 +3,21 @@ import { supabase } from "./supabaseClient.ts"
 
 export type DbTask = Database["public"]["Tables"]["tasks"]["Row"]
 
-export async function addTask(tasks:string):Promise<DbTask | null> {
+export async function getUserId() {
+  const { data: user, error } = await supabase.auth.getUser();
+  if (error) {
+    console.log(error)
+  } else {
+    return user.user.id
+  }
+}
+
+export async function addTask(tasks: string): Promise<DbTask | null> {
+  console.log(tasks)
+  //vi är inte autentiserade här
     const { data, error } = await supabase
         .from("tasks")
-        .insert({ tasks })
+        .insert([{ tasks, is_complete: false, user_id: await getUserId() }])
         .single();
     
     if (error) { 
@@ -54,13 +65,18 @@ export async function updateTaskStatus(id: string, change: boolean | string): Pr
 }
 
  
-export async function deleteTask(id:string):Promise<void> { 
+export async function deleteTask(id: string): Promise<void> { 
   const { error } = await supabase
     .from("tasks")
     .delete()
-    .eq("id", id);
-  if (error) { 
+    .eq("id", id)
+    .eq("user_id", await getUserId()
+    )
+    
+  if (error) {
     console.error("Error deleting task :", error);
+  } else {  
+    fetchTask()
   }
 }
-
+
